+++ Runtime/Extensions/ComparableExtensions.cs
@@ -18,7 +18,7 @@ namespace Zigurous.Math
         /// <param name="includeMin">The minimum value is inclusive if true, exclusive if false.</param>
         /// <param name="includeMax">The maximum value is inclusive if true, exclusive if false.</param>
         /// <returns>True if the value is between the min and max value.</returns>
-        public static bool IsBetween<T>(this T value, T min, T max, bool includeMin = true, bool includeMax = true) where T: IComparable<T>
+        public static bool IsBetween<T>(this T value, T min, T max, bool includeMin = true, bool includeMax = true) where T : IComparable<T>
         {
             int minCompare = value.CompareTo(min);
             int maxCompare = value.CompareTo(max);
@@ -38,7 +38,7 @@ namespace Zigurous.Math
         /// <param name="min">The minimum value.</param>
         /// <param name="max">The maximum value.</param>
         /// <returns>True if the value is in the range [min..max].</returns>
-        public static bool IsBetweenInclusive<T>(this T value, T min, T max) where T: IComparable<T>
+        public static bool IsBetweenInclusive<T>(this T value, T min, T max) where T : IComparable<T>
         {
             return value.IsBetween(min, max, true, true);
         }
@@ -51,7 +51,7 @@ namespace Zigurous.Math
         /// <param name="min">The minimum value.</param>
         /// <param name="max">The maximum value.</param>
         /// <returns>True if the value is in the range [min..max).</returns>
-        public static bool IsBetweenInclusiveExclusive<T>(this T value, T min, T max) where T: IComparable<T>
+        public static bool IsBetweenInclusiveExclusive<T>(this T value, T min, T max) where T : IComparable<T>
         {
             return value.IsBetween(min, max, true, false);
         }
@@ -64,7 +64,7 @@ namespace Zigurous.Math
         /// <param name="min">The minimum value.</param>
         /// <param name="max">The maximum value.</param>
         /// <returns>True if the value is in the range (min..max).</returns>
-        public static bool IsBetweenExclusive<T>(this T value, T min, T max) where T: IComparable<T>
+        public static bool IsBetweenExclusive<T>(this T value, T min, T max) where T : IComparable<T>
         {
             return value.IsBetween(min, max, false, false);
         }
@@ -77,7 +77,7 @@ namespace Zigurous.Math
         /// <param name="min">The minimum value.</param>
         /// <param name="max">The maximum value.</param>
         /// <returns>True if the value is in the range (min..max].</returns>
-        public static bool IsBetweenExclusiveInclusive<T>(this T value, T min, T max) where T: IComparable<T>
+        public static bool IsBetweenExclusiveInclusive<T>(this T value, T min, T max) where T : IComparable<T>
         {
             return value.IsBetween(min, max, false, true);
         }
@@ -88,7 +88,7 @@ namespace Zigurous.Math
         /// <typeparam name="T">The type of value to check.</typeparam>
         /// <param name="values">The values to check.</param>
         /// <returns>The maximum value in the array.</returns>
-        public static T Max<T>(this T[] values) where T: IComparable<T>
+        public static T Max<T>(this T[] values) where T : IComparable<T>
         {
             if (values == null || values.Length == 0) {
                 return default(T);
@@ -114,7 +114,7 @@ namespace Zigurous.Math
         /// <typeparam name="T">The type of value to check.</typeparam>
         /// <param name="values">The values to check.</param>
         /// <returns>The maximum value in the list.</returns>
-        public static T Max<T>(this List<T> values) where T: IComparable<T>
+        public static T Max<T>(this List<T> values) where T : IComparable<T>
         {
             if (values == null || values.Count == 0) {
                 return default(T);
@@ -140,7 +140,7 @@ namespace Zigurous.Math
         /// <typeparam name="T">The type of value to check.</typeparam>
         /// <param name="values">The values to check.</param>
         /// <returns>The minimum value in the array.</returns>
-        public static T Min<T>(this T[] values) where T: IComparable<T>
+        public static T Min<T>(this T[] values) where T : IComparable<T>
         {
             if (values == null || values.Length == 0) {
                 return default(T);
@@ -166,7 +166,7 @@ namespace Zigurous.Math
         /// <typeparam name="T">The type of value to check.</typeparam>
         /// <param name="values">The values to check.</param>
         /// <returns>The minimum value in the list.</returns>
-        public static T Min<T>(this List<T> values) where T: IComparable<T>
+        public static T Min<T>(this List<T> values) where T : IComparable<T>
         {
             if (values == null || values.Count == 0) {
                 return default(T);